# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Atlas Lending intelligence API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Dashboard summaries and activity
  - name: opportunities
    description: Prioritized lending opportunities
  - name: referrals
    description: Referral opportunities
  - name: ingest
    description: Controlled official-source ingestion
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /atlas/dashboard:
    get:
      operationId: getAtlasDashboard
      tags: [dashboard]
      summary: Get the Atlas dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /atlas/orem/source:
    get:
      operationId: getOremSourceReport
      tags: [dashboard]
      summary: Read-only City of Orem July 2026 permit report
      description: Official city PDF evidence, not Utah County-wide coverage or promoted lending opportunities.
      responses:
        "200":
          description: Validated, deduplicated Orem permit signals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OremSourceReport"
        "503":
          description: Official PDF unavailable or failed verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OremSourceError"
  /atlas/opportunities:
    get:
      operationId: listAtlasOpportunities
      tags: [opportunities]
      summary: List prioritized lending opportunities
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: product
          schema:
            type: string
        - in: query
          name: minScore
          schema:
            type: number
            minimum: 0
            maximum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Opportunity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Opportunity"
  /atlas/referrals:
    get:
      operationId: listAtlasReferrals
      tags: [referrals]
      summary: List referral opportunities
      responses:
        "200":
          description: Referral list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Referral"
  /atlas/signals:
    get:
      operationId: listAtlasSignals
      tags: [dashboard]
      summary: List significant signal changes
      responses:
        "200":
          description: Signal changes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SignalChange"
  /atlas/follow-ups:
    get:
      operationId: listAtlasFollowUps
      tags: [dashboard]
      summary: List follow-ups due
      responses:
        "200":
          description: Follow-ups due
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FollowUp"
  /atlas/ingest:
    post:
      operationId: runAtlasIngest
      tags: [ingest]
      summary: Run a controlled official-source ingest
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestInput"
      responses:
        "200":
          description: Ingest result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DashboardSummary:
      type: object
      required: [counts, pipeline, latestIngest, coverage, sourceStatus]
      properties:
        counts:
          $ref: "#/components/schemas/DashboardCounts"
        pipeline:
          type: array
          items:
            $ref: "#/components/schemas/PipelineBucket"
        latestIngest:
          anyOf:
            - $ref: "#/components/schemas/IngestResult"
            - type: "null"
        coverage:
          $ref: "#/components/schemas/CoverageSummary"
        sourceStatus:
          $ref: "#/components/schemas/SourceStatus"
    DashboardCounts:
      type: object
      required: [opportunities, permits, properties, referrals, followUps]
      properties:
        opportunities: { type: integer }
        permits: { type: integer }
        properties: { type: integer }
        referrals: { type: integer }
        followUps: { type: integer }
    PipelineBucket:
      type: object
      required: [label, value, count]
      properties:
        label: { type: string }
        value: { type: number }
        count: { type: integer }
    CoverageSummary:
      type: object
      required: [verified, reviewRequired, rejected]
      properties:
        verified: { type: integer }
        reviewRequired: { type: integer }
        rejected: { type: integer }
    SourceStatus:
      type: object
      required: [state, lastAttemptAt, lastSuccessAt, message]
      properties:
        state:
          type: string
          enum: [CURRENT, STALE, UNAVAILABLE, NEVER]
        lastAttemptAt: { type: ["string", "null"], format: date-time }
        lastSuccessAt: { type: ["string", "null"], format: date-time }
        message: { type: ["string", "null"] }
    OremSourceError:
      type: object
      required: [error]
      properties:
        error: { type: string }
    OremSourceReport:
      type: object
      required: [name, scope, reportThrough, freshness, freshnessMessage, cityPageUrl, monthlyPdfUrl, cumulativePdfUrl, signalCount, propertyCount, properties]
      properties:
        name: { type: string }
        scope: { type: string }
        reportThrough: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
        freshness:
          type: string
          enum: [CURRENT_REPORT, LAGGING_REPORT]
        freshnessMessage: { type: string }
        cityPageUrl: { type: string, format: uri }
        monthlyPdfUrl: { type: string, format: uri }
        cumulativePdfUrl: { type: string, format: uri }
        signalCount: { type: integer }
        propertyCount: { type: integer }
        properties:
          type: array
          items:
            $ref: "#/components/schemas/OremPropertySignals"
    OremPropertySignals:
      type: object
      required: [propertyKey, siteAddress, latestPermitDate, signalCount, totalValuation, score, product, permits]
      properties:
        propertyKey: { type: string }
        siteAddress: { type: string }
        latestPermitDate: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
        signalCount: { type: integer }
        totalValuation: { type: number }
        score: { type: number, minimum: 0, maximum: 1 }
        product:
          type: string
          enum: [UNKNOWN]
        permits:
          type: array
          items:
            $ref: "#/components/schemas/OremPermitSignal"
    OremPermitSignal:
      type: object
      required: [permitId, permitDate, permitType, builder, siteAddress, valuation, score, classification, product, sources]
      properties:
        permitId: { type: string }
        permitDate: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
        permitType: { type: string }
        builder: { type: string }
        siteAddress: { type: string }
        valuation: { type: number }
        score: { type: number, minimum: 0, maximum: 1 }
        classification:
          type: string
          enum: [COMMERCIAL]
        product:
          type: string
          enum: [UNKNOWN]
        sources:
          type: array
          items:
            $ref: "#/components/schemas/OremPdfCitation"
    OremPdfCitation:
      type: object
      required: [url, page, row]
      properties:
        url: { type: string, format: uri }
        page: { type: integer, minimum: 1 }
        row: { type: integer, minimum: 1 }
    Opportunity:
      type: object
      required: [id, company, product, score, status, verificationState, reason, location, updatedAt]
      properties:
        id: { type: integer }
        company: { type: string }
        product: { type: string }
        score: { type: number }
        status: { type: string }
        verificationState: { type: string }
        estimatedMin: { type: ["number", "null"] }
        estimatedMax: { type: ["number", "null"] }
        reason: { type: string }
        location: { type: string }
        propertyType: { type: ["string", "null"] }
        marketValue: { type: ["number", "null"] }
        signalCount: { type: integer }
        updatedAt: { type: string, format: date-time }
    Referral:
      type: object
      required: [id, company, referralType, trigger, score, location, status]
      properties:
        id: { type: integer }
        company: { type: string }
        referralType: { type: string }
        trigger: { type: string }
        score: { type: number }
        location: { type: string }
        status: { type: string }
    SignalChange:
      type: object
      required: [id, label, detail, severity, occurredAt]
      properties:
        id: { type: integer }
        label: { type: string }
        detail: { type: string }
        severity: { type: string }
        occurredAt: { type: string, format: date-time }
    FollowUp:
      type: object
      required: [id, company, task, dueAt, priority, status]
      properties:
        id: { type: integer }
        company: { type: string }
        task: { type: string }
        dueAt: { type: string, format: date-time }
        priority: { type: string }
        status: { type: string }
    IngestInput:
      type: object
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 100
        promoteVerified:
          type: boolean
          default: false
    IngestResult:
      type: object
      required: [runId, seen, written, verifiedMatches, reviewRequired, rejected, promoted]
      properties:
        runId: { type: integer }
        seen: { type: integer }
        written: { type: integer }
        verifiedMatches: { type: integer }
        reviewRequired: { type: integer }
        rejected: { type: integer }
        promoted: { type: boolean }
        startedAt: { type: ["string", "null"], format: date-time }
        finishedAt: { type: ["string", "null"], format: date-time }