# lib/api-zod/src/generated/types/oremPdfCitation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Atlas Lending intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface OremPdfCitation {
  url: string;
  /** @minimum 1 */
  page: number;
  /** @minimum 1 */
  row: number;
}