# lib/api-zod/src/generated/types/oremPermitSignal.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Atlas Lending intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { OremPdfCitation } from './oremPdfCitation';
import type { OremPermitSignalClassification } from './oremPermitSignalClassification';
import type { OremPermitSignalProduct } from './oremPermitSignalProduct';

export interface OremPermitSignal {
  permitId: string;
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  permitDate: string;
  permitType: string;
  builder: string;
  siteAddress: string;
  valuation: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  score: number;
  classification: OremPermitSignalClassification;
  product: OremPermitSignalProduct;
  sources: OremPdfCitation[];
}